hotfix

+++ MaterialMusicDownloader/src/org/upmobile/materialmusicdownloader/adapter/LibraryAdapter.java
@@ -104,6 +104,7 @@ public class LibraryAdapter extends BaseLibraryAdapter implements UndoAdapter, C
 	}
 	
 	public void deleteSong(MusicData song) {
+		if (null == song) return;
 		remove(song);
 		service.remove(song);
 		song.reset(getContext());
